fix(lint): restore swiftlint --strict green on main
Three pre-existing violations were slipping through admin-bypass merges on
recent PRs and CI's lint job has been red on main. Fix them with real
extractions so the baseline is honest again and subsequent Path B
attestations can hold to swiftlint --strict → 0.

- AppModel.swift:111 line_length (158 > 120): split the selectProxy-failed
  error log across a triple-quoted string with `\`-continuations. OSLog
  privacy annotations are preserved because each interpolation still lives
  inside the single logical StringInterpolation literal.

- HomeView.swift:5 type_body_length (302 > 300): move the four `// MARK:
  - Actions` methods (`toggle`, `refreshGroupsIfConnected`, `select`,
  `ping`) into a `private extension HomeView` in the same file. Declarative
  surface stays in the main struct; engine-wiring actions read as a
  separate concern.

- HomeView.swift:609 file_length (609 > 600): extract `ProxyGroupModel`
  (value type + `build(from:)` projector) into a new file at
  `App/Sources/Models/ProxyGroupModel.swift`. It's already decoupled from
  SwiftUI — only consumes `Proxy` from MeowModels — so the move is pure
  code motion.

Thresholds in `.swiftlint.yml` stay at the prior 600/300/800/400 values.
